# specs/001-/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: 攝影作品展示網站 API
  description: 支援年份、作品集與照片管理的 Next.js API Routes
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API Routes
components:
  schemas:
...
    Collection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        year_id:
          type: string
          format: uuid
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        title:
          type: string
          maxLength: 200
        summary:
          type: string
          maxLength: 500
        cover_asset_id:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published]
        order_index:
          type: string
        published_at:
          type: string
          format: date-time
          nullable: true
      required: [id, year_id, slug, title, status, order_index]
    Asset:
      type: object
      properties:
        id:
          type: string
          description: "Cloudflare Images image_id"
        alt:
          type: string
          minLength: 1
          maxLength: 200
        caption:
          type: string
          maxLength: 1000
          nullable: true
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        metadata_json:
          type: object
          nullable: true
      required: [id, alt, width, height]
...
    DirectUploadResponse:
      type: object
      properties:
        upload_url:
          type: string
          format: uri
        image_id:
          type: string
        form_data:
          type: object
          additionalProperties: true
      required: [upload_url, image_id]
...
  securitySchemes:
    CloudflareAccess:
      type: http
      scheme: bearer
      description: "Cloudflare Access JWT token"
paths:
...
  /years/{year_id}/collections:
    get:
      summary: 獲取年份下的作品集列表
      parameters:
        - name: year_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, all]
          example: published
      responses:
        '200':
          description: 成功獲取作品集列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collection'

    post:
      summary: 在年份下創建新作品集
      security:
        - CloudflareAccess: []
      parameters:
        - name: year_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slug:
                  type: string
                title:
                  type: string
                summary:
                  type: string
                cover_asset_id:
                  type: string
                  nullable: true
                status:
                  type: string
                  enum: [draft, published]
                  default: draft
                order_index:
                  type: string
              required: [slug, title]
      responses:
        '201':
          description: 成功創建作品集
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
  /collections/{collection_id}:
    get:
      summary: 獲取作品集詳細資訊
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: include_assets
          in: query
          schema:
            type: boolean
            default: false
          description: "是否包含作品集中的照片列表"
      responses:
        '200':
          description: 成功獲取作品集資訊
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Collection'
                  - type: object
                    properties:
                      assets:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Asset'
                            - type: object
                              properties:
                                order_index:
                                  type: string

    put:
      summary: 更新作品集資訊
      security:
        - CloudflareAccess: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                summary:
                  type: string
                cover_asset_id:
                  type: string
                  nullable: true
                status:
                  type: string
                  enum: [draft, published]
                order_index:
                  type: string
      responses:
        '200':
          description: 成功更新作品集
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'

    delete:
      summary: 刪除作品集
      security:
        - CloudflareAccess: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功刪除作品集

  # 照片與資產管理 API
  /images/direct-upload:
    post:
      summary: 獲取 Cloudflare Images 直傳憑證
      security:
        - CloudflareAccess: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                content_type:
                  type: string
                  example: "image/jpeg"
              required: [filename]
      responses:
        '200':
          description: 成功獲取上傳憑證
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectUploadResponse'
        '400':
          description: 不支援的檔案類型
...
  /revalidate:
    post:
      summary: 觸發快取重新驗證
      security:
        - CloudflareAccess: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
                  description: "要失效的快取標籤"
                paths:
                  type: array
                  items:
                    type: string
                  description: "要重新驗證的路徑"
              required: [tags]
      responses:
        '200':
          description: 成功觸發重新驗證
          content:
            application/json:
              schema:
                type: object
                properties:
                  revalidated:
                    type: array
                    items:
                      type: string